✅(backend) fix lyra tests
After a rebase, as tests have been split, we need to rework them a bit.
