+++ src/backend/joanie/tests/payment/test_backend_lyra.py
@@ -1986,81 +1986,3 @@ class LyraBackendTestCase(BasePaymentTestCase, LoggingTestCase):
                 },
             },
         )
-
-    @patch.object(BasePaymentBackend, "_do_on_batch_order_payment_failure")
-    def test_payment_backend_lyra_handle_notification_payment_failure_for_batch_order(
-        self, mock_do_on_batch_order_payment_failure
-    ):
-        """
-        When backend receives a payment notification which failed for a batch order, the generic
-        method `_do_on_payment_failure_for_batch_order` should be called.
-        """
-        backend = LyraBackend(self.configuration)
-        batch_order = BatchOrderFactory(
-            state=BATCH_ORDER_STATE_PENDING,
-            id="758c2570-a7af-4335-b091-340d0cc6e694",
-            owner__email="john.doe@acme.org",
-            relation__product__price=D("123.45"),
-        )
-
-        with self.open("lyra/requests/payment_refused.json") as file:
-            json_request = json.loads(file.read())
-
-        request = APIRequestFactory().post(
-            reverse("payment_webhook"), data=json_request, format="multipart"
-        )
-
-        backend.handle_notification(request)
-
-        mock_do_on_batch_order_payment_failure.assert_called_once_with(
-            batch_order=batch_order,
-        )
-
-    @patch.object(BasePaymentBackend, "_do_on_batch_order_payment_success")
-    def test_payment_backend_lyra_handle_notification_for_batch_order(
-        self, mock_do_on_batch_order_payment_success
-    ):
-        """
-        When we receive a notification from the successful payment of a batch order,
-        the generic method `_do_on_batch_order_payment_success` should be called.
-        """
-        backend = LyraBackend(self.configuration)
-        batch_order = BatchOrderFactory(
-            id="514070fe-c12c-48b8-97cf-5262708673a3",
-            owner__email="john.doe@acme.org",
-            state=BATCH_ORDER_STATE_PENDING,
-            relation__product__price=D("123.45"),
-            nb_seats=1,
-        )
-
-        with self.open("lyra/requests/payment_accepted_no_store_card.json") as file:
-            json_request = json.loads(file.read())
-
-        with self.open(
-            "lyra/requests/payment_accepted_no_store_card_answer.json"
-        ) as file:
-            json_answer = json.loads(file.read())
-
-        request = APIRequestFactory().post(
-            reverse("payment_webhook"), data=json_request, format="multipart"
-        )
-
-        backend.handle_notification(request)
-
-        transaction_id = json_answer["transactions"][0]["uuid"]
-        billing_details = json_answer["customer"]["billingDetails"]
-        mock_do_on_batch_order_payment_success.assert_called_once_with(
-            batch_order=batch_order,
-            payment={
-                "id": transaction_id,
-                "amount": D("123.45"),
-                "billing_address": {
-                    "address": billing_details["address"],
-                    "city": billing_details["city"],
-                    "country": billing_details["country"],
-                    "first_name": billing_details["firstName"],
-                    "last_name": billing_details["lastName"],
-                    "postcode": billing_details["zipCode"],
-                },
-            },
-        )
